Commit_06_Refactoring
In this commit  I made major logical changes to the structure of the classes. Moved methods "AimAt" and "Fire" from class Tank to class Aiming Component (Now class Tanks is only used to derive a Tank Blueprint). Code is now much easier to read and understand, also all of the pointers are protected.